Updating folder structure
