+++ src/js/colors/index.ts
@@ -1,0 +1,2 @@
+export { colors as default } from './colors';
+export { colors } from './colors';
