+++ src/js/index.ts
@@ -1,0 +1,2 @@
+export { colors } from './colors';
+export { typography } from './typography';
