+++ src/js/typography/index.ts
@@ -1,0 +1,1 @@
+export { typography } from './typography';
