--- src/ts/typography/index.ts
@@ -1,6 +1,0 @@
-export { fontFamilies } from './fontFamilies';
-export { fontSizes } from './fontSizes';
-export { fontWeights } from './fontWeights';
-export { lineHeights } from './lineHeights';
-export { letterSpacing } from './letterSpacing';
-export { typography } from './typography';
